app/requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
psycopg2==2.9.6
python-dotenv==1.0.0
requests==2.29.0
urllib3==1.26.15
